fix: 适配器模式改造

+++ src/main/java/com/yupi/springbootinit/service/impl/PostServiceImpl.java
@@ -307,8 +307,8 @@ public class PostServiceImpl extends ServiceImpl<PostMapper, Post> implements Po
 
     @Override
     public Page<PostVO> listPostVOByPage(PostQueryRequest postQueryRequest,HttpServletRequest request) {
-        int current = postQueryRequest.getCurrent();
-        int pageSize = postQueryRequest.getPageSize();
+        long current = postQueryRequest.getCurrent();
+        long pageSize = postQueryRequest.getPageSize();
         Page<Post> postPage = this.page(new Page<>(current, pageSize),
                 this.getQueryWrapper(postQueryRequest));
         return this.getPostVOPage(postPage,request);
